add flux capacitor
